Move _plugin.actions flat into _plugin
