+++ extended_mypy_django_plugin/_plugin/sem_analyze.py
@@ -23,11 +23,9 @@ from mypy.types import (
     UnionType,
     get_proper_type,
 )
-from mypy.types import (
-    Type as MypyType,
-)
+from mypy.types import Type as MypyType
 
-from .. import protocols
+from . import protocols
 
 
 class TypeAnalyzer:
