+++ extended_mypy_django_plugin/_plugin/type_checker.py
@@ -42,7 +42,7 @@ from mypy.types import (
 from mypy.types import Type as MypyType
 from typing_extensions import Self
 
-from .. import protocols
+from . import protocols
 
 
 class ResolveManagerMethodFromInstance(Protocol):
